Modularizes test results components
Moves test results-related components to a new dedicated file for better separation of concerns and reusability.
Introduces type definitions for test results to enhance type safety and maintainability.
